requirements.txt
google-cloud-aiplatform
google-cloud-pipeline-components==2.9.0
kfp
pytest